+++ sigstore/_cli.py
@@ -222,10 +222,10 @@ def _add_shared_oidc_options(
         help="The OpenID Connect issuer to use (conflicts with --staging)",
     )
     group.add_argument(
-        "--oidc-disable-default-browser",
+        "--oauth-force-oob",
         action="store_true",
-        default=_boolify_env("SIGSTORE_OIDC_DISABLE_DEFAULT_BROWSER"),
-        help="Do not start the default web browser to complete the OAuth flow. Instead display a web link to the user.",
+        default=_boolify_env("SIGSTORE_OAUTH_FORCE_OOB"),
+        help="Force an out-of-band OAuth flow and do not automatically start the default web browser",
     )
 
 
@@ -972,7 +972,7 @@ def _get_identity(args: argparse.Namespace) -> Optional[IdentityToken]:
     token = issuer.identity_token(
         client_id=args.oidc_client_id,
         client_secret=args.oidc_client_secret,
-        force_oob=args.oidc_disable_default_browser,
+        force_oob=args.oath_force_oob,
     )
 
     return token